Implement basic comparison of band structures
This commit implements a basic comparison of band structures in the
SBSPlotter class. If the class is passed a list of BandStructureSymmLine
objects, it will attempt to plot them on the same plot, assuming they
have compatible k-paths (similar behavior to the base BSPlotter in
pymatgen).

This commit further factors the color and line style
selection out of the get_plot method.